Rename DataSyncClientPlugin to OffixDataSyncPlugin
